+++ packages/bootstrap-vue-3/src/components/BCard/BCardImg.vue
@@ -4,30 +4,39 @@
 
 <script setup lang="ts">
 // import type {BCardImgProps} from '../../types/components'
+import type {Booleanish} from '../../types'
+import {resolveBooleanish} from '../../utils'
 import {computed} from 'vue'
 
 interface BCardImgProps {
   alt?: string
-  bottom?: boolean
-  end?: boolean
+  bottom?: Booleanish
+  end?: Booleanish
   height?: number | string
-  left?: boolean
-  right?: boolean
+  left?: Booleanish
+  right?: Booleanish
   src?: string
-  start?: boolean
-  top?: boolean
+  start?: Booleanish
+  top?: Booleanish
   width?: number | string
 }
 
 const props = withDefaults(defineProps<BCardImgProps>(), {
-  alt: undefined,
   bottom: false,
   end: false,
   left: false,
   right: false,
+  alt: undefined,
   start: false,
   top: false,
 })
+
+const bottomBoolean = computed<boolean>(() => resolveBooleanish(props.bottom))
+const endBoolean = computed<boolean>(() => resolveBooleanish(props.end))
+const leftBoolean = computed<boolean>(() => resolveBooleanish(props.left))
+const rightBoolean = computed<boolean>(() => resolveBooleanish(props.right))
+const startBoolean = computed<boolean>(() => resolveBooleanish(props.start))
+const topBoolean = computed<boolean>(() => resolveBooleanish(props.top))
 
 const attrs = computed(() => ({
   src: props.src,
@@ -41,17 +50,17 @@ const attrs = computed(() => ({
 }))
 
 const classes = computed(() => {
-  const align = props.left ? 'float-left' : props.right ? 'float-right' : ''
+  const align = leftBoolean.value ? 'float-left' : rightBoolean.value ? 'float-right' : ''
 
   let baseClass = 'card-img'
 
-  if (props.top) {
+  if (topBoolean.value) {
     baseClass += '-top'
-  } else if (props.right || props.end) {
+  } else if (rightBoolean.value || endBoolean.value) {
     baseClass += '-right'
-  } else if (props.bottom) {
+  } else if (bottomBoolean.value) {
     baseClass += '-bottom'
-  } else if (props.left || props.start) {
+  } else if (leftBoolean.value || startBoolean.value) {
     baseClass += '-left'
   }