+++ packages/bootstrap-vue-3/src/components/BFormCheckbox/index.ts
@@ -1,3 +1,0 @@
-import BFormCheckbox from './BFormCheckbox.vue'
-
-export default BFormCheckbox