+++ packages/bootstrap-vue-3/src/components/BPopover.vue
@@ -32,10 +32,11 @@ import {
   ref,
   watch,
 } from 'vue'
+import {resolveBooleanish} from '../utils'
 import Popover from 'bootstrap/js/dist/popover'
 import {useEventListener} from '../composables'
 import type {BPopoverDelayObject} from '../types/components'
-import type {ColorVariant} from '../types'
+import {Booleanish, ColorVariant} from '../types'
 
 export default defineComponent({
   props: {
@@ -47,7 +48,7 @@ export default defineComponent({
     },
     content: {type: String},
     id: {type: String},
-    noninteractive: {type: Boolean, default: false},
+    noninteractive: {type: Boolean as PropType<Booleanish>, default: false},
     placement: {type: String as PropType<Popover.Options['placement']>, default: 'right'},
     target: {
       type: [String, Object] as PropType<
@@ -58,14 +59,21 @@ export default defineComponent({
     title: {type: String},
     delay: {type: [Number, Object] as PropType<number | BPopoverDelayObject>, default: 0},
     triggers: {type: String as PropType<Popover.Options['trigger']>, default: 'click'},
-    show: {type: Boolean, default: false},
+    show: {type: Boolean as PropType<Booleanish>, default: false},
     variant: {type: String as PropType<ColorVariant>, default: undefined},
-    html: {type: Boolean, default: true},
-    sanitize: {type: Boolean, default: false},
+    html: {type: Boolean as PropType<Booleanish>, default: true},
+    sanitize: {type: Boolean as PropType<Booleanish>, default: false},
     offset: {type: String as PropType<Popover.Options['offset']>, default: '0'},
   },
   emits: ['show', 'shown', 'hide', 'hidden', 'inserted'],
   setup(props, {emit, slots}) {
+    // TODO noninteractive is never used
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const noninteractiveBoolean = computed<boolean>(() => resolveBooleanish(props.noninteractive))
+    const showBoolean = computed<boolean>(() => resolveBooleanish(props.show))
+    const htmlBoolean = computed<boolean>(() => resolveBooleanish(props.html))
+    const sanitizeBoolean = computed<boolean>(() => resolveBooleanish(props.sanitize))
+
     const element = ref<HTMLElement>()
     const target = ref<HTMLElement | undefined>()
     const instance = ref<Popover>()
@@ -108,9 +116,9 @@ export default defineComponent({
         placement: props.placement,
         title: props.title || slots.title ? titleRef.value : '',
         content: contentRef.value,
-        html: props.html,
+        html: htmlBoolean.value,
         delay: props.delay,
-        sanitize: props.sanitize,
+        sanitize: sanitizeBoolean.value,
         offset: props.offset,
       })
     }
@@ -122,7 +130,7 @@ export default defineComponent({
 
       element.value?.parentNode?.removeChild(element.value)
 
-      if (props.show) {
+      if (showBoolean.value) {
         instance.value?.show()
       }
     })
@@ -140,7 +148,7 @@ export default defineComponent({
     )
 
     watch(
-      () => props.show,
+      () => showBoolean.value,
       (show, oldVal) => {
         if (show !== oldVal) {
           if (show) {
@@ -166,157 +174,4 @@ export default defineComponent({
     }
   },
 })
-/*
-// import type {BPopoverEmits, BPopoverProps} from '../types/components'
-import type {BPopoverDelayObject} from '../types/components'
-import Popover from 'bootstrap/js/dist/popover'
-import {
-  ComponentPublicInstance,
-  computed,
-  defineComponent,
-  nextTick,
-  onBeforeUnmount,
-  onMounted,
-  ref,
-  useSlots,
-  watch,
-} from 'vue'
-import {useEventListener} from '../composables'
-import type {ColorVariant} from '../types'
-
-interface BPopoverProps {
-  id?: string
-  title?: string
-  content?: string
-  container?: string | ComponentPublicInstance<HTMLElement> | HTMLElement
-  noninteractive?: boolean
-  placement?: Popover.Options['placement']
-  target?: string | ComponentPublicInstance<HTMLElement> | HTMLElement
-  delay?: number | BPopoverDelayObject
-  triggers?: Popover.Options['trigger']
-  show?: boolean
-  variant?: ColorVariant
-  html?: boolean
-  sanitize?: boolean
-  offset?: Popover.Options['offset']
-}
-
-const props = withDefaults(defineProps<BPopoverProps>(), {
-  container: 'body',
-  noninteractive: false,
-  placement: 'right',
-  target: undefined,
-  delay: 0,
-  triggers: 'click',
-  show: false,
-  variant: undefined,
-  html: true,
-  sanitize: false,
-  offset: '0',
-})
-
-interface BPopoverEmits {
-  (e: 'show'): void
-  (e: 'shown'): void
-  (e: 'hide'): void
-  (e: 'hidden'): void
-  (e: 'inserted'): void
-}
-
-const emit = defineEmits<BPopoverEmits>()
-
-const slots = useSlots()
-
-const element = ref<HTMLElement>()
-const target = ref<HTMLElement | undefined>()
-const instance = ref<Popover>()
-const titleRef = ref<HTMLElement>()
-const contentRef = ref<HTMLElement>()
-const classes = computed(() => ({
-  [`b-popover-${props.variant}`]: props.variant,
-}))
-
-const cleanElementProp = (
-  target: string | ComponentPublicInstance<HTMLElement> | HTMLElement | undefined
-): HTMLElement | string | undefined => {
-  if (typeof target === 'string') {
-    return target
-  } else if (target instanceof HTMLElement) return target
-  else if (typeof target !== 'undefined')
-    return (target as ComponentPublicInstance<HTMLElement>).$el as HTMLElement
-  return undefined
-}
-
-const getElement = (element: HTMLElement | string | undefined): HTMLElement | undefined => {
-  if (!element) return undefined
-  if (typeof element === 'string') {
-    const idElement = document.getElementById(element)
-    return idElement ? idElement : undefined
-  }
-  return element
-}
-
-const generatePopoverInstance = (
-  targetValue: string | ComponentPublicInstance<HTMLElement> | HTMLElement | undefined
-) => {
-  target.value = getElement(cleanElementProp(targetValue))
-
-  if (!target.value) return
-
-  instance.value = new Popover(target.value, {
-    container: cleanElementProp(props.container),
-    trigger: props.triggers,
-    placement: props.placement,
-    title: props.title || slots.title ? titleRef.value : '',
-    content: contentRef.value,
-    html: props.html,
-    delay: props.delay,
-    sanitize: props.sanitize,
-    offset: props.offset,
-  })
-}
-
-onMounted(() => {
-  nextTick(() => {
-    generatePopoverInstance(props.target)
-  })
-
-  element.value?.parentNode?.removeChild(element.value)
-
-  if (props.show) {
-    instance.value?.show()
-  }
-})
-
-onBeforeUnmount(() => {
-  instance.value?.dispose()
-})
-
-watch(
-  () => props.target,
-  (newValue) => {
-    instance.value?.dispose()
-    generatePopoverInstance(newValue)
-  }
-)
-
-watch(
-  () => props.show,
-  (show, oldVal) => {
-    if (show !== oldVal) {
-      if (show) {
-        instance.value?.show()
-      } else {
-        instance.value?.hide()
-      }
-    }
-  }
-)
-
-useEventListener(target, 'show.bs.popover', () => emit('show'))
-useEventListener(target, 'shown.bs.popover', () => emit('shown'))
-useEventListener(target, 'hide.bs.popover', () => emit('hide'))
-useEventListener(target, 'hidden.bs.popover', () => emit('hidden'))
-useEventListener(target, 'inserted.bs.popover', () => emit('inserted'))
-*/
 </script>