+++ packages/bootstrap-vue-3/src/components/BTable/BTbody.vue
@@ -6,17 +6,24 @@
 
 <script setup lang="ts">
 // import type {BTBodyProps} from '../../types/components'
+import type {Booleanish} from '../../types'
+import {resolveBooleanish} from '../../utils'
 import {computed} from 'vue'
 
 interface BTBodyProps {
-  headVariant?: boolean
+  headVariant?: Booleanish
 }
 
 const props = withDefaults(defineProps<BTBodyProps>(), {
   headVariant: false,
 })
 
+const headVariantBoolean = computed<boolean>(() => resolveBooleanish(props.headVariant))
+
 const classes = computed(() => ({
-  [`thead-${props.headVariant}`]: props.headVariant,
+  // TODO this doesn't look right thead-false?
+  // This also goes for BTfoot
+  // BThead is string, not boolean. That is probably correct
+  [`thead-${props.headVariant}`]: headVariantBoolean.value,
 }))
 </script>